+++ test/unit/fuzz-tools.test.js
@@ -289,6 +289,7 @@ import * as gvhdV309 from '../../lib/gvhd-v309.js';
 import * as cholangitisV310 from '../../lib/cholangitis-v310.js';
 import * as cholecystitisV311 from '../../lib/cholecystitis-v311.js';
 import * as cholangitisDxV312 from '../../lib/cholangitis-dx-v312.js';
+import * as cholecystitisDxV313 from '../../lib/cholecystitis-dx-v313.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -523,6 +524,7 @@ const MODULES = {
   'cholangitis-v310.js': cholangitisV310,
   'cholecystitis-v311.js': cholecystitisV311,
   'cholangitis-dx-v312.js': cholangitisDxV312,
+  'cholecystitis-dx-v313.js': cholecystitisDxV313,
   'tb-testing.js': tbTesting,
 };